Unifico todos los campos de fechas de la base de datos a tipo DATE (en reservas y recordatorios) y obtengo todos los mensajes mostrados al usuario directamente desde el backend
Con esos objetivos:
 - Realizo las modificaciones necesarias para evitar problemas relacionados a zonas horarias en el almacenamiento de las fechas
 - Traduzco los mensajes devueltos por la API del backend a español, para mostrarle dichos mensajes directamente al usuario en el frontend

Además:

BACKEND

 - Elimino imágenes de la carpeta '/public' que no se correspondían con el formato en el que actualmente se almacenan las imágenes en dicha carpeta
 - Agrego una validación adicional a la hora de realizar una nueva reserva, para que antes de registrarla vuelva a confirmar que el vehículo sigue disponible (en caso contrario se ejecuta un rollback y se le muestra el mensaje correspondiente de error al usuario en el frontend)
- Separo la lógica del Check-Out de una reserva de la de un simple update, para poder manejar correctamente la actualización, tanto de la reserva como del vehículo, dentro de una transacción
 - Modifico la lógica de eliminación de las reservas, para que solo sea posible en el caso de que todavía no se le haya mandado el recordatorio al cliente
 - Recorto espacios exteriores en el valor recibido en los métodos de verificación de unicidad antes de comparar, para mejorar la validación
 - Convierto el documentID de los usuarios a mayúsculas antes de guardarlo, para unificar el formato de los pasaportes que contienen letras
 - Solo devuelvo mensaje de 'Error al eliminar la imagen' de un modelo si no se trata de que la misma no fue encontrada, para mejorar la experiencia del usuario al no mostrarle un error innecesario

FRONTEND

 - Centralizo la apertura del modal genérico GenericDialogComponent a un servicio compartido (OpenDialogService), cuyos métodos utilizan el resto de los componentes
 - Unifico la forma de mostrar los mensajes de error para que los mismos se muestren a través del modal genérico (salvo en casos específicos), redireccionando a distintas rutas al momento del cierre del modal según corresponda
 - Modularizo el código TypeScript de los componentes en varias funciones separadas, para que, por ejemplo, no se lleve a cabo todo en el ngOnInit( ), onSubmit( ), entre otras
 - Defino funciones específicas 'handleSubmit( )', 'handleError( )' y algunas otras de nombres similares, para que manejen las distintas respuestas provenientes del backend
 - Le agrego el modificador de acceso 'private' a aquellos métodos que sólo son accedidos desde el propio código TypeScript del componente en el que están definidos
 - Paralelizo solicitudes para obtener información del backend en ReservationFormComponent y ReservationStepperComponent, para mejorar el rendimiento
 - Elimino código redundante en el archivo HTML de EmailVerificationComponent, manejado ahora con los mensajes dinámicos devueltos por el backend
 - Renombro los métodos 'clickEvent( )' utilizados para modificar la visibilidad de los campos relacionados a contraseñas, a 'togglePasswordVisibility( )'
 - Elimino dicho método y las variables e importaciones correspondientes en el ProfileComponent, dado que, al no tener su formulario un campo de contraseña, no se utilizaban
 - Elimino archivo 'date-formats.config.ts' ya no utilizado
 - Modifico el tamaño de algunas de las imágenes guardadas en '/assets' dentro del frontend, para que sean más acordes al tamaño en el que son mostradas, y así evitar posibles errores

GENERAL

 - Modifico la imagen del modelo de datos, para que sea acorde a los cambios mencionados sobre los campos de fechas y al agregado del campo final_price en la tabla de reservas realizado en el commit anterior
 - Modifico algunos otros pequeños detalles en el README.md principal

+++ backend/src/core/reservation/reservation.controller.ts
@@ -2,11 +2,15 @@
 import { Request, Response } from 'express';
 
 // MikroORM
+import { EntityManager } from '@mikro-orm/core';
 import { orm } from '../../shared/database/orm.js';
 
 // Entities
 import { Reservation } from './reservation.entity.js';
 import { Reminder } from '../reminder/reminder.entity.js';
+
+// External Libraries
+import { subDays } from 'date-fns';
 
 const em = orm.em;
 
@@ -20,31 +24,46 @@ const findAll = async (_req: Request, res: Response) => {
       }
     );
     res.status(200).json({
-      message: 'All reservations have been found',
+      message: 'Todas las reservas han sido encontradas',
       data: reservations,
     });
   } catch (error) {
-    res.status(500).json({ message: 'Server error' });
+    res.status(500).json({ message: 'Error de conexión' });
   }
 };
 
 const add = async (req: Request, res: Response) => {
+  const orm = em.fork();
+  await orm.begin();
+
   try {
-    const reservation = em.create(Reservation, req.body.sanitizedInput);
+    const reservationInput = req.body.sanitizedInput;
+    const { startDate, plannedEndDate, vehicle } = reservationInput;
 
-    const reminderDate = new Date(req.body.startDate);
-    reminderDate.setHours(reminderDate.getHours() + 3);
-    reminderDate.setDate(reminderDate.getDate() - 1);
+    const existingReservation = await validateReservation(
+      orm,
+      vehicle,
+      startDate,
+      plannedEndDate
+    );
 
-    em.create(Reminder, {
-      reminderDate,
-      sent: false,
-      reservation,
-    });
-    await em.flush();
-    res.status(201).end();
+    if (existingReservation) {
+      await orm.rollback();
+      return res.status(400).json({
+        message:
+          'El vehículo ya no está disponible para las fechas seleccionadas. Por favor, seleccione otro vehículo o intente con otro rango de fechas.',
+      });
+    }
+
+    await createReservation(orm, reservationInput, reservationInput.user);
+
+    await orm.flush();
+    await orm.commit();
+
+    res.status(201).json({ message: 'Reserva exitosa' });
   } catch (error) {
-    res.status(500).json({ message: 'Server error' });
+    await orm.rollback();
+    res.status(500).json({ message: 'Error de conexión' });
   }
 };
 
@@ -53,37 +72,79 @@ const update = async (req: Request, res: Response) => {
     const id = Number.parseInt(req.params.id);
     const reservation = await em.findOne(Reservation, { id });
     if (!reservation) {
-      return res
-        .status(404)
-        .json({ message: 'The reservation does not exist' });
+      return res.status(404).json({ message: 'Reserva no encontrada' });
     } else {
       em.assign(reservation, req.body.sanitizedInput);
       await em.flush();
-      res.status(200).json({ message: 'The reservation has been updated' });
+      res
+        .status(200)
+        .json({ message: 'La reserva ha sido actualizada exitosamente' });
     }
   } catch (error) {
-    res.status(500).json({ message: 'Server error' });
+    res.status(500).json({ message: 'Error de conexión' });
+  }
+};
+
+const checkOut = async (req: Request, res: Response) => {
+  const orm = em.fork();
+  await orm.begin();
+
+  try {
+    const id = Number.parseInt(req.params.id);
+
+    const reservation = await orm.findOne(
+      Reservation,
+      { id },
+      { populate: ['vehicle'] }
+    );
+
+    if (!reservation) {
+      await orm.rollback();
+      return res.status(404).json({ message: 'Reserva no encontrada' });
+    }
+
+    const vehicle = reservation.vehicle;
+
+    orm.assign(vehicle, { totalKms: req.body.sanitizedInput.finalKms });
+    orm.assign(reservation, req.body.sanitizedInput);
+
+    await orm.flush();
+    await orm.commit();
+
+    res.status(200).json({
+      message: 'Se ha realizado el check-out de la reserva exitosamente',
+    });
+  } catch (error) {
+    await orm.rollback();
+    res.status(500).json({ message: 'Error de conexión' });
   }
 };
 
 const remove = async (req: Request, res: Response) => {
   try {
     const id = Number.parseInt(req.params.id);
-    const reservation = await em.findOne(Reservation, { id });
-    const reservationInUse = await em.findOne(Reminder, { reservation: id });
+    const reservation = await em.findOne(
+      Reservation,
+      { id },
+      { populate: ['reminders'] }
+    );
     if (!reservation) {
-      res.status(404).json({ message: 'Reservation does not exist' });
-    } else if (reservationInUse) {
-      res.status(400).json({
-        message:
-          'La reserva no se puede eliminar porque tiene recordatorios asociados.',
-      });
-    } else {
-      await em.removeAndFlush(reservation);
-      res.status(200).json({ message: 'The reservation has been deleted' });
+      return res.status(404).json({ message: 'Reserva no encontrada' });
     }
+    const reminders = reservation.reminders.getItems();
+    const hasSentReminders = reminders.some((reminder) => reminder.sent);
+    if (hasSentReminders) {
+      return res.status(400).json({
+        message:
+          'La reserva no se puede eliminar porque ya se ha enviado un recordatorio al cliente.',
+      });
+    }
+    await em.removeAndFlush(reservation);
+    res
+      .status(200)
+      .json({ message: 'La reserva ha sido eliminada exitosamente' });
   } catch (error) {
-    res.status(500).json({ message: 'Server error' });
+    res.status(500).json({ message: 'Error de conexión' });
   }
 };
 
@@ -100,36 +161,91 @@ const getReservationsByUser = async (req: Request, res: Response) => {
       }
     );
     res.status(200).json({
-      message: 'All reservations have been found',
+      message: 'Todas sus reservas han sido encontradas',
       data: reservationsByUser,
     });
   } catch (error) {
-    res.status(500).json({ message: 'Server error' });
+    res.status(500).json({ message: 'Error de conexión' });
   }
 };
 
 const userReservation = async (req: Request, res: Response) => {
+  const orm = em.fork();
+  await orm.begin();
+
   try {
-    const reservation = em.create(Reservation, {
-      ...req.body.sanitizedInput,
-      user: req.session.user.id,
-    });
+    const reservationInput = req.body.sanitizedInput;
+    const { startDate, plannedEndDate, vehicle } = reservationInput;
 
-    const reminderDate = new Date(req.body.startDate);
-    reminderDate.setHours(reminderDate.getHours() + 3);
-    reminderDate.setDate(reminderDate.getDate() - 1);
+    const existingReservation = await validateReservation(
+      orm,
+      vehicle,
+      startDate,
+      plannedEndDate
+    );
 
-    em.create(Reminder, {
-      reminderDate,
-      sent: false,
-      reservation,
-    });
+    if (existingReservation) {
+      await orm.rollback();
+      return res.status(400).json({
+        message:
+          'El vehículo ya no está disponible para las fechas seleccionadas. Por favor, seleccione otro vehículo o intente con otro rango de fechas.',
+      });
+    }
 
-    await em.flush();
-    res.status(201).end();
+    await createReservation(orm, reservationInput, req.session.user.id);
+
+    await orm.flush();
+    await orm.commit();
+
+    res.status(201).json({ message: 'Reserva exitosa' });
   } catch (error) {
-    res.status(500).json({ message: 'Server error' });
+    await orm.rollback();
+    res.status(500).json({ message: 'Error de conexión' });
   }
 };
 
-export { findAll, add, update, remove, getReservationsByUser, userReservation };
+const validateReservation = async (
+  em: EntityManager,
+  vehicle: number,
+  startDate: string,
+  plannedEndDate: string
+) => {
+  return em.findOne(Reservation, {
+    vehicle,
+    realEndDate: null,
+    cancellationDate: null,
+    startDate: { $lte: plannedEndDate },
+    plannedEndDate: { $gte: startDate },
+  });
+};
+
+const createReservation = async (
+  em: EntityManager,
+  reservationInput: any,
+  userId: number
+) => {
+  const reservation = em.create(Reservation, {
+    ...reservationInput,
+    user: userId,
+  });
+  em.create(Reminder, {
+    reminderDate: getReminderDate(reservation.startDate),
+    sent: false,
+    reservation,
+  });
+  return reservation;
+};
+
+const getReminderDate = (reservationStartDate: string) => {
+  return subDays(reservationStartDate, 1).toISOString().split('T')[0];
+};
+
+export {
+  findAll,
+  add,
+  update,
+  checkOut,
+  remove,
+  getReservationsByUser,
+  userReservation,
+};
